Add files via upload
The Weather program converts Celsius temperatures to Fahrenheit and prints both sets of temperatures. It first initializes an array of temperatures in Celsius, then calls a method to convert them to Fahrenheit. The converted temperatures are printed alongside the original Celsius temperatures. Additionally, the program includes a method to print temperatures, which can be called for both Celsius and Fahrenheit arrays.